+++ packages/expo/src/launch/RootErrorBoundary.tsx
@@ -42,6 +42,7 @@ export default class RootErrorBoundary extends React.Component<Props, State> {
   _subscribeToGlobalErrors = () => {
     this._appLoadingIsMounted = true;
 
+    // Bacon: This isn't supported in RNWeb yet
     let ErrorUtils = global.ErrorUtils;
 
     if (!ErrorUtils) return;
