+++ packages/expo/src/launch/registerRootComponent.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { AppRegistry } from 'react-native';
-import withExpoRoot, { InitialProps } from './withExpoRoot';
+import withExpoRoot from './withExpoRoot';
+import { InitialProps } from './withExpoRoot.types';
 
 export default function registerRootComponent<P extends InitialProps>(
   component: React.ComponentClass<P>
