+++ packages/expo/src/launch/registerRootComponent.web.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import ReactDOM from 'react-dom';
-import withExpoRoot, { InitialProps } from './withExpoRoot';
+import withExpoRoot from './withExpoRoot';
+import { InitialProps } from './withExpoRoot.types';
 
 export default function registerRootComponent<P extends InitialProps>(
   component: React.ComponentClass<P>
