+++ packages/expo/src/launch/withExpoRoot.types.ts
@@ -1,0 +1,10 @@
+export type InitialProps = {
+  exp: {
+    notification?: any;
+    errorRecovery?: any;
+    [key: string]: any;
+  };
+  shell?: boolean;
+  shellManifestUrl?: string;
+  [key: string]: any;
+};
